2023_H2/cos/3/ocp/requirements.txt
pyqt5==5.15.9
Pillow
numpy